cleanup several filenames and move a few around
