+++ packages/ado-npm-auth/src/npmrc/check-tokens.ts
@@ -1,14 +1,15 @@
-import { makeADORequest } from "./makeADORequest.js";
-import { getUserPat } from "./getUserPAT.js";
+import { makeADORequest } from "./make-ado-request.js";
+import { getUserPat } from "./get-user-pat.js";
 import { NpmrcOrg } from "./get-repo-npmrc-ado-orgs.js";
 import { getUserNPMRC } from "./npmrc.js";
 
 /**
+ * Validates the tokens for each feed/organization by making a simple network request to ADO
  * @param { Object } options
  * @param { boolean } [options.silent]
  * @param { Array<NpmrcOrg> } [options.feeds]
  */
-export const checkADO = async function ({ feeds }: { feeds: NpmrcOrg[] }) {
+export const checkTokens = async function ({ feeds }: { feeds: NpmrcOrg[] }) {
   const userNpmRc = getUserNPMRC();
   
   const feedsWithPat = await getUserPat({ npmrc: userNpmRc, feeds });
@@ -27,8 +28,7 @@ export const checkADO = async function ({ feeds }: { feeds: NpmrcOrg[] }) {
       });
     }
 
-  } catch (/** @type {any} */ e) {
-
+  } catch (e) {
     return false;
   }
