+++ packages/ado-npm-auth/src/utils/get-feed-without-protocol.ts
