+++ packages/node-azureauth/src/azure-auth-command.ts
